Move project creation and opening to projectController. Fixes #48 Fixes #52
